fix use different visibility check
